requirements.txt
clickhouse-driver==0.2.6
pandas==2.2.3
kafka-python-ng>=1.0.0